Create a generic solver using ortools CPSat solver
We generalize what has been made for `CPSatRCPSPSolver` by creating an
abstract class `OrtoolsCPSatSolver`.

It needs to implement `retrieve_solution` which takes an ortools
callback `cpsolvercb` as input, that can be used to get the variables
value for the current solution via `cpsolvercb.Value(VARIABLE_NAME)`.

This ortools callback is called at each new solution found by the cp-sat solver to
- update the result_storage
- call the users (do) callbacks, that can used for logging but also
  early stopping if return True.
