+++ discrete_optimization/rcpsp/solver/cpsat_solver.py
@@ -8,10 +8,6 @@ import logging
 from typing import Any, Dict, Hashable, List, Optional, Set, Tuple
 
 from ortools.sat.python.cp_model import (
-    FEASIBLE,
-    INFEASIBLE,
-    OPTIMAL,
-    UNKNOWN,
     CpModel,
     CpSolver,
     CpSolverSolutionCallback,
@@ -19,20 +15,9 @@ from ortools.sat.python.cp_model import (
     IntVar,
 )
 
-from discrete_optimization.generic_tools.callbacks.callback import (
-    Callback,
-    CallbackList,
-)
-from discrete_optimization.generic_tools.cp_tools import (
-    CPSolver,
-    ParametersCP,
-    StatusSolver,
-)
+from discrete_optimization.generic_tools.cp_tools import StatusSolver
 from discrete_optimization.generic_tools.do_problem import ParamsObjectiveFunction
-from discrete_optimization.generic_tools.exceptions import SolveEarlyStop
-from discrete_optimization.generic_tools.result_storage.result_storage import (
-    ResultStorage,
-)
+from discrete_optimization.generic_tools.ortools_cpsat_tools import OrtoolsCPSatSolver
 from discrete_optimization.rcpsp.rcpsp_model import RCPSPModel, RCPSPSolution
 from discrete_optimization.rcpsp.rcpsp_utils import (
     create_fake_tasks,
@@ -45,7 +30,7 @@ from discrete_optimization.rcpsp.special_constraints import PairModeConstraint
 logger = logging.getLogger(__name__)
 
 
-class CPSatRCPSPSolver(CPSolver, SolverRCPSP):
+class CPSatRCPSPSolver(OrtoolsCPSatSolver, SolverRCPSP):
     def __init__(
         self,
         problem: RCPSPModel,
@@ -241,6 +226,7 @@ class CPSatRCPSPSolver(CPSolver, SolverRCPSP):
                 model.Add(score_task[task1] == score_task[task2])
 
     def init_model(self, **kwargs):
+        """Init CP model."""
         include_special_constraints = kwargs.get(
             "include_special_constraints", self.problem.includes_special_constraint()
         )
@@ -286,37 +272,19 @@ class CPSatRCPSPSolver(CPSolver, SolverRCPSP):
             "is_present": is_present_var,
         }
 
-    def solve(
-        self,
-        callbacks: Optional[List[Callback]] = None,
-        parameters_cp: Optional[ParametersCP] = None,
-        **kwargs: Any,
-    ) -> ResultStorage:
-        callbacks_list = CallbackList(callbacks=callbacks)
-        callbacks_list.on_solve_start(solver=self)
-        if self.cp_model is None:
-            self.init_model(**kwargs)
-        if parameters_cp is None:
-            parameters_cp = ParametersCP.default()
-        solver = CpSolver()
-        solver.parameters.max_time_in_seconds = parameters_cp.time_limit
-        solver.parameters.num_workers = parameters_cp.nb_process
-        ortools_callback = OrtoolsCallback(do_solver=self, callback=callbacks_list)
-        try:
-            status = solver.Solve(self.cp_model, ortools_callback)
-            self.status_solver = cpstatus_to_dostatus(status_from_cpsat=status)
-        except SolveEarlyStop as e:
-            logger.info(e)
-            if ortools_callback.nb_solutions > 0:
-                status = StatusSolver.SATISFIED
-            else:
-                status = StatusSolver.UNSATISFIABLE
-            self.status_solver = cpstatus_to_dostatus(status_from_cpsat=status)
-        res = ortools_callback.res
-        callbacks_list.on_solve_end(res=res, solver=self)
-        return res
-
     def retrieve_solution(self, cpsolvercb: CpSolverSolutionCallback) -> RCPSPSolution:
+        """Construct a do solution from the cpsat solver internal solution.
+
+        It will be called each time the cpsat solver find a new solution.
+        At that point, value of internal variables are accessible via `cpsolvercb.Value(VARIABLE_NAME)`.
+
+        Args:
+            cpsolvercb: the ortools callback called when the cpsat solver finds a new solution.
+
+        Returns:
+            the intermediate solution, at do format.
+
+        """
         schedule = {}
         modes_dict = {}
         for task in self.variables["start"]:
@@ -615,49 +583,3 @@ class CPSatRCPSPSolverCumulativeResource(CPSatRCPSPSolver):
             "is_present": is_present_var,
             "resource_capacity": resource_capacity_var,
         }
-
-
-def cpstatus_to_dostatus(status_from_cpsat) -> StatusSolver:
-    """
-
-    :param status_from_cpsat: either [UNKNOWN,INFEASIBLE,OPTIMAL,FEASIBLE] from ortools.cp api.
-    :return: Status
-    """
-    if status_from_cpsat == UNKNOWN:
-        return StatusSolver.UNKNOWN
-    if status_from_cpsat == INFEASIBLE:
-        return StatusSolver.UNSATISFIABLE
-    if status_from_cpsat == OPTIMAL:
-        return StatusSolver.OPTIMAL
-    if status_from_cpsat == FEASIBLE:
-        return StatusSolver.SATISFIED
-
-
-class OrtoolsCallback(CpSolverSolutionCallback):
-    def __init__(self, do_solver: CPSatRCPSPSolver, callback: Callback):
-        super().__init__()
-        self.do_solver = do_solver
-        self.callback = callback
-        self.res = ResultStorage(
-            [],
-            mode_optim=self.do_solver.params_objective_function.sense_function,
-            limit_store=False,
-        )
-        self.nb_solutions = 0
-
-    def on_solution_callback(self) -> None:
-        self.store_current_solution()
-        self.nb_solutions += 1
-        # end of step callback: stopping?
-        stopping = self.callback.on_step_end(
-            step=self.nb_solutions, res=self.res, solver=self.do_solver
-        )
-        if stopping:
-            raise SolveEarlyStop(
-                f"{self.do_solver.__class__.__name__}.solve() stopped by user callback."
-            )
-
-    def store_current_solution(self):
-        sol = self.do_solver.retrieve_solution(cpsolvercb=self)
-        fit = self.do_solver.aggreg_from_sol(sol)
-        self.res.add_solution(solution=sol, fitness=fit)
